implement login and register fragment
