+++ app/src/main/java/com/example/criminalintentapp/presentation/MainActivity.kt
@@ -6,29 +6,10 @@ import com.example.criminalintentapp.R
 import com.example.criminalintentapp.presentation.fragments.crime_detail.CrimeFragment
 import com.example.criminalintentapp.presentation.fragments.crime_list.CrimeListFragment
 
-class MainActivity : AppCompatActivity(), CrimeListFragment.Callbacks {
+class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-     /*  val currentFragment = supportFragmentManager.findFragmentById(R.id.fragment_container)
-
-        if (currentFragment == null) {
-            val fragment = CrimeListFragment.newInstance()
-            supportFragmentManager
-                .beginTransaction()
-                .add(R.id.fragment_container, fragment)
-                .commit()
-        }*/
-    }
-
-    override fun onCrimeSelected(crimeId: Int) {
-       /* val fragment = CrimeFragment.newInstance(crimeId)
-        supportFragmentManager
-            .beginTransaction()
-            .replace(R.id.fragment_container, fragment)
-            .addToBackStack(null)
-            .commit() */
     }
 }
