+++ app/src/main/res/navigation/navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.example.criminalintentapp.LoginFragment"
+        android:name="com.example.criminalintentapp.presentation.authentication.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -15,11 +15,13 @@
             app:destination="@id/registerFragment" />
         <action
             android:id="@+id/action_loginFragment_to_crimeListFragment"
-            app:destination="@id/crimeListFragment" />
+            app:destination="@id/crimeListFragment"
+            app:popUpTo="@id/navigation"
+            app:popUpToInclusive="true" />
     </fragment>
     <fragment
         android:id="@+id/registerFragment"
-        android:name="com.example.criminalintentapp.RegisterFragment"
+        android:name="com.example.criminalintentapp.presentation.authentication.RegisterFragment"
         android:label="fragment_register"
         tools:layout="@layout/fragment_register" >
         <action
@@ -27,7 +29,9 @@
             app:destination="@id/loginFragment" />
         <action
             android:id="@+id/action_registerFragment_to_crimeListFragment"
-            app:destination="@id/crimeListFragment" />
+            app:destination="@id/crimeListFragment"
+            app:popUpTo="@id/navigation"
+            app:popUpToInclusive="true" />
     </fragment>
     <fragment
         android:id="@+id/crimeListFragment"
@@ -36,6 +40,11 @@
         <action
             android:id="@+id/action_crimeListFragment_to_crimeFragment"
             app:destination="@id/crimeFragment" />
+        <action
+            android:id="@+id/action_crimeListFragment_to_loginFragment"
+            app:destination="@id/loginFragment"
+            app:popUpTo="@id/navigation"
+            app:popUpToInclusive="true" />
     </fragment>
     <fragment
         android:id="@+id/crimeFragment"
@@ -43,6 +52,8 @@
         android:label="CrimeFragment" >
         <action
             android:id="@+id/action_crimeFragment_to_crimeListFragment"
-            app:destination="@id/crimeListFragment" />
+            app:destination="@id/crimeListFragment"
+            app:popUpTo="@id/navigation"
+            app:popUpToInclusive="true" />
     </fragment>
 </navigation>
